+++ src/definitions/types/buff/buff.ts
@@ -1,4 +1,4 @@
-import type { DamageBuffRestrictedTo } from '../../../models/v4/buff/damage-buff-restricted-to';
+import type { DamageBuffRestrictedTo } from '../../../models/v4/buff/damage-buff/damage-buff-restricted-to';
 
 export interface Buff {
   value: number;
