+++ src/models/v4/buff/attack-buff/attack-buff-aggregate.ts
@@ -1,5 +1,5 @@
-import type { WeaponElementalType } from '../../../definitions/elemental-type';
-import { sum } from '../../../utils/math-utils';
+import type { WeaponElementalType } from '../../../../definitions/elemental-type';
+import { sum } from '../../../../utils/math-utils';
 import type { AttackBuff } from './attack-buff';
 
 export class AttackBuffAggregate {
