+++ src/models/v4/buff/base-attack-buff/base-attack-buff-aggregate.ts
@@ -1,5 +1,5 @@
-import type { WeaponElementalType } from '../../../definitions/elemental-type';
-import { sum } from '../../../utils/math-utils';
+import type { WeaponElementalType } from '../../../../definitions/elemental-type';
+import { sum } from '../../../../utils/math-utils';
 import type { BaseAttackBuff } from './base-attack-buff';
 
 export class BaseAttackBuffAggregate {
