+++ src/models/v4/buff/buff-event.ts
@@ -6,12 +6,12 @@ import type { EventManager } from '../event/event-manager';
 import type { CurrentTick } from '../tick/current-tick';
 import type { TimeInterval } from '../time-interval/time-interval';
 import type { ActiveBuff } from './active-buff/active-buff';
-import type { AttackBuff } from './attack-buff';
-import type { BaseAttackBuff } from './base-attack-buff';
-import type { CritDamageBuff } from './crit-damage-buff';
+import type { AttackBuff } from './attack-buff/attack-buff';
+import type { BaseAttackBuff } from './base-attack-buff/base-attack-buff';
+import type { CritDamageBuff } from './crit-damage-buff/crit-damage-buff';
 import type { BuffEventDto } from './dtos/buff-event-dto';
-import type { ElementalDamageBuff } from './elemental-damage-buff';
-import type { FinalDamageBuff } from './final-damage-buff';
+import type { ElementalDamageBuff } from './elemental-damage-buff/elemental-damage-buff';
+import type { FinalDamageBuff } from './final-damage-buff/final-damage-buff';
 import type { MiscellaneousBuff } from './miscellaneous-buff';
 
 export class BuffEvent
