+++ src/models/v4/buff/crit-damage-buff/crit-damage-buff.ts
@@ -1,4 +1,4 @@
-import { Buff } from './buff';
+import { Buff } from '../buff';
 
 export class CritDamageBuff extends Buff {
   public override canApplyTo(): boolean {
