+++ src/models/v4/buff/crit-rate-buff/crit-rate-buff-aggregate.ts
@@ -1,4 +1,4 @@
-import { sum } from '../../../utils/math-utils';
+import { sum } from '../../../../utils/math-utils';
 import type { CritRateBuff } from './crit-rate-buff';
 
 export class CritRateBuffAggregate {
