+++ src/models/v4/buff/crit-rate-buff/crit-rate-buff.ts
@@ -1,4 +1,4 @@
-import { Buff } from './buff';
+import { Buff } from '../buff';
 
 export class CritRateBuff extends Buff {
   public override canApplyTo(): boolean {
