+++ src/models/v4/buff/damage-buff/damage-buff-restricted-to.ts
@@ -1,6 +1,6 @@
-import type { AttackType } from '../../../definitions/attack-type';
-import type { AttackId } from '../../../definitions/types/attack/attack-ability';
-import type { WeaponName } from '../../../definitions/weapons/weapon-definitions';
+import type { AttackType } from '../../../../definitions/attack-type';
+import type { AttackId } from '../../../../definitions/types/attack/attack-ability';
+import type { WeaponName } from '../../../../definitions/weapons/weapon-definitions';
 
 export interface DamageBuffRestrictedTo {
   /** Damage buff only applies to attacks of this weapon */
