+++ src/models/v4/buff/damage-buff/damage-buff.ts
@@ -1,7 +1,7 @@
-import type { BuffId } from '../../../definitions/types/buff/buff-ability';
-import type { AttackHit } from '../event/messages/attack-hit';
-import { Buff } from './buff';
-import type { BuffSource } from './buff-source';
+import type { BuffId } from '../../../../definitions/types/buff/buff-ability';
+import type { AttackHit } from '../../event/messages/attack-hit';
+import { Buff } from '../buff';
+import type { BuffSource } from '../buff-source';
 import type { DamageBuffRestrictedTo } from './damage-buff-restricted-to';
 
 export abstract class DamageBuff extends Buff {
