+++ src/models/v4/buff/elemental-damage-buff/elemental-damage-buff.ts
@@ -1,9 +1,9 @@
-import type { WeaponElementalType } from '../../../definitions/elemental-type';
-import type { BuffId } from '../../../definitions/types/buff/buff-ability';
-import type { AttackHit } from '../event/messages/attack-hit';
-import type { BuffSource } from './buff-source';
-import { DamageBuff } from './damage-buff';
-import type { DamageBuffRestrictedTo } from './damage-buff-restricted-to';
+import type { WeaponElementalType } from '../../../../definitions/elemental-type';
+import type { BuffId } from '../../../../definitions/types/buff/buff-ability';
+import type { AttackHit } from '../../event/messages/attack-hit';
+import type { BuffSource } from '../buff-source';
+import { DamageBuff } from '../damage-buff/damage-buff';
+import type { DamageBuffRestrictedTo } from '../damage-buff/damage-buff-restricted-to';
 
 export class ElementalDamageBuff extends DamageBuff {
   public constructor(
