--- src/models/v4/buff/final-damage-buff.ts
@@ -1,3 +1,0 @@
-import { DamageBuff } from './damage-buff';
-
-export class FinalDamageBuff extends DamageBuff {}
