+++ src/models/v4/buff/final-damage-buff/final-damage-buff.ts
@@ -1,0 +1,3 @@
+import { DamageBuff } from '../damage-buff/damage-buff';
+
+export class FinalDamageBuff extends DamageBuff {}
