+++ src/models/v4/combat-simulator/combat-simulator.ts
@@ -21,16 +21,16 @@ import { AttackAbility } from '../attack/attack-ability';
 import { AttackRegistry } from '../attack/attack-registry';
 import { AttackTimeline } from '../attack/attack-timeline';
 import { ActiveBuffs } from '../buff/active-buff/active-buffs';
-import { AttackBuff } from '../buff/attack-buff';
-import { BaseAttackBuff } from '../buff/base-attack-buff';
+import { AttackBuff } from '../buff/attack-buff/attack-buff';
+import { BaseAttackBuff } from '../buff/base-attack-buff/base-attack-buff';
 import { BuffAbility } from '../buff/buff-ability';
 import { BuffRegistry } from '../buff/buff-registry';
 import type { BuffSource } from '../buff/buff-source';
 import { BuffTimeline } from '../buff/buff-timeline';
-import { CritDamageBuff } from '../buff/crit-damage-buff';
-import { CritRateBuff } from '../buff/crit-rate-buff';
-import { ElementalDamageBuff } from '../buff/elemental-damage-buff';
-import { FinalDamageBuff } from '../buff/final-damage-buff';
+import { CritDamageBuff } from '../buff/crit-damage-buff/crit-damage-buff';
+import { CritRateBuff } from '../buff/crit-rate-buff/crit-rate-buff';
+import { ElementalDamageBuff } from '../buff/elemental-damage-buff/elemental-damage-buff';
+import { FinalDamageBuff } from '../buff/final-damage-buff/final-damage-buff';
 import { UtilizedBuffs } from '../buff/utilized-buffs';
 import { CurrentCharacterStats } from '../character/current-character-stats';
 import { Charge } from '../charge/charge';
