+++ src/models/v4/damage-record/damage-record.ts
@@ -5,8 +5,8 @@ import { product, sum } from '../../../utils/math-utils';
 import type { Serializable } from '../../persistable';
 import type { ActiveBuffs } from '../buff/active-buff/active-buffs';
 import type { Buff } from '../buff/buff';
-import { ElementalDamageBuffAggregate } from '../buff/elemental-damage-buff-aggregate';
-import { FinalDamageBuffAggregate } from '../buff/final-damage-buff-aggregate';
+import { ElementalDamageBuffAggregate } from '../buff/elemental-damage-buff/elemental-damage-buff-aggregate';
+import { FinalDamageBuffAggregate } from '../buff/final-damage-buff/final-damage-buff-aggregate';
 import type { UtilizedBuffs } from '../buff/utilized-buffs';
 import type { CurrentCharacterStats } from '../character/current-character-stats';
 import { Damage } from '../damage/damage';
